split field_widgets.py into option_display.py, cleanup class
