+++ nixui/graphics/field_widgets.py
@@ -6,201 +6,6 @@ from nixui.options import api, option_tree, option_definition, types
 from nixui.graphics import color_indicator, richtext, generic_widgets, toggle_switch
 
 
-def get_field_widget_classes_from_type(option_type):
-    if isinstance(option_type, types.ListOfType):
-        return [ListOfRedirect]
-    elif isinstance(option_type, types.AttrsOfType):
-        return [AttrsOfRedirect]
-    elif isinstance(option_type, types.AttrsType):
-        return [AttrsRedirect]
-    elif isinstance(option_type, types.SubmoduleType):
-        return [SubmoduleRedirect]
-    elif isinstance(option_type, types.EitherType):
-        widgets = set()
-        for subtype in option_type.subtypes:
-            widgets |= set(get_field_widget_classes_from_type(subtype))
-        return list(widgets)
-    elif isinstance(option_type, types.UnspecifiedType):
-        return [UndefinedField]
-    elif isinstance(option_type, types.NullType):
-        return [NullField]
-    elif isinstance(option_type, types.BoolType):
-        return [BooleanField]
-    elif isinstance(option_type, types.StrType):
-        return [TextField]
-    elif isinstance(option_type, types.IntType):
-        return [IntegerField]
-    elif isinstance(option_type, types.OneOfType):
-        return [OneOfField]
-    elif isinstance(option_type, types.PathType):
-        return [NotImplementedField]
-    elif isinstance(option_type, types.PackageType):
-        return [NotImplementedField]
-    elif isinstance(option_type, types.FunctionType):
-        return [NotImplementedField]
-    elif isinstance(option_type, types.AnythingType):
-        return [NotImplementedField]
-    else:
-        raise NotImplementedError(option_type)
-
-
-def get_field_widget_classes(option_type):
-    return (
-        [UndefinedField] +
-        get_field_widget_classes_from_type(option_type) +
-        [ExpressionField, ReferenceField]
-    )
-
-
-def get_label_color_for_widget(field_widget):
-    field_colors = {
-        UndefinedField: QtGui.QColor(255, 200, 200),  # TODO: create
-        ExpressionField: QtGui.QColor(193, 236, 245),
-        ReferenceField: QtGui.QColor(174, 250, 174),
-    }
-    return field_colors.get(
-        field_widget,
-        QtGui.QColor(255, 255, 240),  # default
-    )
-
-
-class GenericOptionDisplay(QtWidgets.QWidget):
-    def __init__(self, statemodel, set_option_path_fn, option, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        self.statemodel = statemodel
-        self.set_option_path_fn = set_option_path_fn
-
-        self.option = option
-        self.starting_definition = None
-
-        field_widget_classes = get_field_widget_classes(
-            types.from_nix_type_str(
-                api.get_option_tree().get_type(option)
-            )
-        )
-
-        # set fields for entry editing
-        self.entry_stack = QtWidgets.QStackedWidget()
-        for field_widget_class in field_widget_classes:
-            entry_widget = field_widget_class(self.option)
-            # TODO: fix this hacky handling of `Redirect` (#109)
-            if not isinstance(entry_widget, Redirect):
-                entry_widget.stateChanged.connect(self.handle_state_change)
-                self.statemodel.slotmapper.add_slot(('update_field', self.option), self._load_definition)
-            self.entry_stack.addWidget(entry_widget)
-        self.stacked_widgets = list(map(self.entry_stack.widget, range(self.entry_stack.count())))
-
-        # set type selector
-        self.field_type_selector = generic_widgets.ExclusiveButtonGroup(
-            choices=[
-                (
-                    w.name,
-                    self.set_type,
-                    get_label_color_for_widget(w)
-                )
-                for w in self.stacked_widgets
-            ]
-        )
-        # TODO: remove this when reference editor is done
-        self.field_type_selector.btn_group.buttons()[-1].setEnabled(False)
-
-        # set title and description
-        text = generic_widgets.ClickableLabel(str(option))
-        text.clicked.connect(lambda: set_option_path_fn(option))
-        tooltip = generic_widgets.ToolTip(
-            richtext.get_option_html(
-                option,
-                type_label=api.get_option_tree().get_type(option),
-                description=api.get_option_tree().get_description(option),
-            )
-        )
-
-        # add all to layout
-        option_details_layout = QtWidgets.QHBoxLayout()
-        option_details_layout.addWidget(tooltip)
-        option_details_layout.addWidget(text)
-        option_details_layout.addStretch()
-        description_layout = QtWidgets.QVBoxLayout()
-        description_layout.addLayout(option_details_layout)
-        description_layout.addWidget(self.field_type_selector, stretch=1)
-        layout = QtWidgets.QHBoxLayout()
-        layout.addLayout(description_layout)
-        layout.addWidget(self.entry_stack, stretch=8)
-        self.setLayout(layout)
-
-        self._load_definition()
-
-    def _load_definition(self):
-        option_definition = self.statemodel.get_definition(self.option)
-
-        for i, field in enumerate(self.stacked_widgets):
-            if field.validate_field(option_definition.obj):
-                self.field_type_selector.select(i)
-                field.load_value(option_definition.obj)
-                break
-        else:
-            self.field_type_selector.select(len(self.stacked_widgets) - 1)
-            expression_field = self.stacked_widgets[-2]
-            expression_field.load_value(option_definition.expression_string)
-
-        self.starting_definition = self.definition
-
-    def set_type(self, arg):
-        stack_idx = self.field_type_selector.checked_index()
-        current_widget = self.entry_stack.widget(stack_idx)
-        definition = self.statemodel.get_definition(self.option)
-        # TODO: fix this hacky handling of `Redirect` (#109)
-        if isinstance(current_widget, Redirect):
-            self.set_option_path_fn(self.option, current_widget.option_type)
-            return
-        elif isinstance(current_widget, ExpressionField):
-            current_widget.load_value(definition.expression_string)
-        else:
-            current_widget.load_value(definition.obj)
-        self.entry_stack.setCurrentIndex(stack_idx)
-        self.handle_state_change()
-
-    def handle_state_change(self):
-        self.statemodel.slotmapper('form_definition_changed')(self.option, self.definition)
-
-    @property
-    def definition(self):
-        current_widget = self.entry_stack.currentWidget()
-        form_value = current_widget.current_value
-        if isinstance(current_widget, ExpressionField):
-            return option_definition.OptionDefinition.from_expression_string(form_value)
-        else:
-            return option_definition.OptionDefinition.from_object(form_value)
-
-    def contains_focus(self):
-        return (
-            self.hasFocus() or
-            any(w.hasFocus() for w in self.stacked_widgets) or
-            any(w.hasFocus() for w in self.field_type_selector.btn_group.buttons())
-        )
-
-    def paint_background_color(self, bg_color):
-        qp = QtGui.QPainter(self)
-        r = QtCore.QRect(0, 0, self.width(), self.height())
-        qp.fillRect(r, bg_color)
-        qp.end()
-
-    def paintEvent(self, ev):
-        super().paintEvent(ev)
-        if self.contains_focus():
-            self.paint_background_color(
-                QtGui.QColor(233, 245, 248)
-            )
-        else:
-            bg_color = color_indicator.get_edit_state_color_indicator(
-                api.get_option_tree(),
-                self.option
-            )
-            self.paint_background_color(bg_color)
-        self.update()
-
-
 # TODO: fix this hacky handling of `Redirect` (#109)
 class Redirect(QtWidgets.QLabel):
     """
