+++ nixui/graphics/option_display.py
@@ -1,0 +1,232 @@
+from nixui.options import api, types, option_definition
+from nixui.graphics import color_indicator, richtext, generic_widgets, field_widgets, toggle_switch
+
+from PyQt5 import QtWidgets, QtGui, QtCore
+
+
+def get_field_widget_classes_from_type(option_type):
+    if isinstance(option_type, types.ListOfType):
+        return [field_widgets.ListOfRedirect]
+    elif isinstance(option_type, types.AttrsOfType):
+        return [field_widgets.AttrsOfRedirect]
+    elif isinstance(option_type, types.AttrsType):
+        return [field_widgets.AttrsRedirect]
+    elif isinstance(option_type, types.SubmoduleType):
+        return [field_widgets.SubmoduleRedirect]
+    elif isinstance(option_type, types.EitherType):
+        widgets = set()
+        for subtype in option_type.subtypes:
+            widgets |= set(get_field_widget_classes_from_type(subtype))
+        return list(widgets)
+    elif isinstance(option_type, types.UnspecifiedType):
+        return [field_widgets.UndefinedField]
+    elif isinstance(option_type, types.NullType):
+        return [field_widgets.NullField]
+    elif isinstance(option_type, types.BoolType):
+        return [field_widgets.BooleanField]
+    elif isinstance(option_type, types.StrType):
+        return [field_widgets.TextField]
+    elif isinstance(option_type, types.IntType):
+        return [field_widgets.IntegerField]
+    elif isinstance(option_type, types.OneOfType):
+        return [field_widgets.OneOfField]
+    elif isinstance(option_type, types.PathType):
+        return [field_widgets.NotImplementedField]
+    elif isinstance(option_type, types.PackageType):
+        return [field_widgets.NotImplementedField]
+    elif isinstance(option_type, types.FunctionType):
+        return [field_widgets.NotImplementedField]
+    elif isinstance(option_type, types.AnythingType):
+        return [field_widgets.NotImplementedField]
+    else:
+        raise NotImplementedError(option_type)
+
+
+def get_field_widget_classes(option_type):
+    return (
+        [field_widgets.UndefinedField] +
+        get_field_widget_classes_from_type(option_type) +
+        [field_widgets.ExpressionField, field_widgets.ReferenceField]
+    )
+
+
+def get_label_color_for_widget(field_widget):
+    field_colors = {
+        field_widgets.UndefinedField: QtGui.QColor(255, 200, 200),  # TODO: create
+        field_widgets.ExpressionField: QtGui.QColor(193, 236, 245),
+        field_widgets.ReferenceField: QtGui.QColor(174, 250, 174),
+    }
+    return field_colors.get(
+        field_widget,
+        QtGui.QColor(255, 255, 240),  # default
+    )
+
+
+class GenericOptionDisplay(QtWidgets.QWidget):
+    def __init__(self, statemodel, set_option_path_fn, option, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.statemodel = statemodel
+        self.set_option_path_fn = set_option_path_fn
+        self.option = option
+
+        # vbox containing read-only option details and "is-defined" toggle switch
+        description_layout = QtWidgets.QVBoxLayout()
+        description_layout.addLayout(
+            self._get_option_details_layout(option, set_option_path_fn)
+        )
+        self.is_defined_toggle = toggle_switch.ToggleSwitch("Defined", "Undefined")
+        self.is_defined_toggle.stateChanged.connect(self.set_is_defined)
+        description_layout.addWidget(self.is_defined_toggle)
+
+        # field widget selector and stacked widget containing field widgets for display
+        definition_layout = QtWidgets.QHBoxLayout()
+        self.entry_stack = self._get_field_stack_widget(option)
+        self.stacked_widgets = list(map(self.entry_stack.widget, range(self.entry_stack.count())))
+        self.field_selector = self._get_field_selection_layout(self.stacked_widgets)
+        definition_layout.addWidget(self.field_selector)
+        definition_layout.addWidget(self.entry_stack)
+
+        # stack to toggle visibility in definition
+        self.definition_stack = QtWidgets.QStackedWidget()
+        self.definition_stack.addWidget(QtWidgets.QLabel())
+        self.definition_stack.addLayout(definition_layout)
+
+        # put together horizontally
+        layout = QtWidgets.QHBoxLayout()
+        layout.addLayout(description_layout)
+        layout.addWidget(self.definition_stack)
+        self.setLayout(layout)
+
+        self._load_definition()
+
+    @staticmethod
+    def _get_option_details_layout(option, set_option_path_fn):
+        # title and description
+        text = generic_widgets.ClickableLabel(str(option))
+        text.clicked.connect(lambda: set_option_path_fn(option))
+        tooltip = generic_widgets.ToolTip(
+            richtext.get_option_html(
+                option,
+                type_label=api.get_option_tree().get_type(option),
+                description=api.get_option_tree().get_description(option),
+            )
+        )
+        option_details_layout = QtWidgets.QHBoxLayout()
+        option_details_layout.addWidget(tooltip)
+        option_details_layout.addWidget(text)
+        option_details_layout.addStretch()
+        return option_details_layout
+
+    def _get_field_stack_widget(self, option):
+        field_widget_classes = get_field_widget_classes(
+            types.from_nix_type_str(
+                api.get_option_tree().get_type(option)
+            )
+        )
+        entry_stack = QtWidgets.QStackedWidget()
+        for field_widget_class in field_widget_classes:
+            entry_widget = field_widget_class(option)
+            # TODO: fix this hacky handling of `Redirect` (#109)
+            if not isinstance(entry_widget, field_widgets.Redirect):
+                entry_widget.stateChanged.connect(self.handle_state_change)
+                self.statemodel.slotmapper.add_slot(('update_field', option), self._load_definition)
+            entry_stack.addWidget(entry_widget)
+
+        return entry_stack
+
+    def _get_field_selection_layout(self, field_widgets):
+        # TODO: remove this when reference editor is done
+        #self.field_selector.btn_group.buttons()[-1].setEnabled(False)
+
+        return generic_widgets.ExclusiveButtonGroup(
+            choices=[
+                (
+                    w.name,
+                    self.set_type,
+                    get_label_color_for_widget(w)
+                )
+                for w in field_widgets
+            ]
+        )
+
+    def _load_definition(self):
+        option_definition = self.statemodel.get_definition(self.option)
+
+        if option_definition.is_undefined:
+            self.is_defined_toggle.setChecked(False)
+            return
+        else:
+            self.is_defined_toggle.setChecked(True)
+
+        for i, field in enumerate(self.stacked_widgets):
+            if field.validate_field(option_definition.obj):
+                self.field_selector.select(i)
+                field.load_value(option_definition.obj)
+                break
+        else:
+            self.field_selector.select(len(self.stacked_widgets) - 1)
+            expression_field = self.stacked_widgets[-2]
+            expression_field.load_value(option_definition.expression_string)
+
+    def set_type(self, arg=None):
+        stack_idx = self.field_selector.checked_index()
+        current_widget = self.entry_stack.widget(stack_idx)
+        definition = self.statemodel.get_definition(self.option)
+        # TODO: fix this hacky handling of `Redirect` (#109)
+        if isinstance(current_widget, field_widgets.Redirect):
+            self.set_option_path_fn(self.option, current_widget.option_type)
+            return
+        elif isinstance(current_widget, field_widgets.ExpressionField):
+            current_widget.load_value(definition.expression_string)
+        else:
+            current_widget.load_value(definition.obj)
+        self.entry_stack.setCurrentIndex(stack_idx)
+        self.handle_state_change()
+
+    def set_is_defined(self):
+        if self.is_defined_toggle.isChecked():
+            self.set_type()
+            self.definition_stack.setCurrentIndex(1)
+        else:
+            self.definition_stack.setCurrentIndex(0)
+        self.handle_state_change()
+
+
+    def handle_state_change(self):
+        self.statemodel.slotmapper('form_definition_changed')(self.option, self.definition)
+
+    @property
+    def definition(self):
+        current_widget = self.entry_stack.currentWidget()
+        form_value = current_widget.current_value
+        if isinstance(current_widget, field_widgets.ExpressionField):
+            return option_definition.OptionDefinition.from_expression_string(form_value)
+        else:
+            return option_definition.OptionDefinition.from_object(form_value)
+
+    def contains_focus(self):
+        return (
+            self.hasFocus() or
+            any(w.hasFocus() for w in self.stacked_widgets) or
+            any(w.hasFocus() for w in self.field_selector.btn_group.buttons())
+        )
+
+    def paint_background_color(self, bg_color):
+        qp = QtGui.QPainter(self)
+        r = QtCore.QRect(0, 0, self.width(), self.height())
+        qp.fillRect(r, bg_color)
+        qp.end()
+
+    def paintEvent(self, ev):
+        super().paintEvent(ev)
+        if self.contains_focus():
+            self.paint_background_color(
+                QtGui.QColor(233, 245, 248)
+            )
+        else:
+            bg_color = color_indicator.get_edit_state_color_indicator(
+                api.get_option_tree(),
+                self.option
+            )
+            self.paint_background_color(bg_color)
+        self.update()
